WebVR: Moved setProjectionFromUnion method to internal WebVRUtils.
